clean up routes and controllers
