+++ app/Controllers/Http/UserController.js
@@ -217,50 +217,6 @@ class UserController {
 		return response.redirect('/');
 	}
 
-	/**
-	 * Log a user in and redirect them to their respective landing page depending on the user type.
-	 *
-	 * @param {Object} Context The context object.
-	 */
-	async login ({ request, auth, response, session }) {
-		const { email } = request.all();
-		const user = await User
-			.query()
-			.where('email', email.toLowerCase())
-			.where('verified', true)
-			.first();
-
-		try {
-			await auth.login(user);
-			if (auth.user.getUserRole() === 'User') {
-				session.flash({
-					notification: 'Welcome! You are logged in'
-				});
-
-				return response.redirect('/userDash');
-			} else {
-				return response.redirect('/');
-			}
-		} catch (error) {
-			session.flash({ loginError: 'Invalid email/password' });
-			return response.redirect('/login');
-		}
-	}
-
-	/**
-	 * Log a user out.
-	 *
-	 * @param {Object} Context The context object.
-	 */
-	async logout ({ auth, response, session }) {
-		await auth.logout();
-
-		session.flash({
-			notification: 'You have been logged out.'
-		});
-		return response.redirect(`${Env.get('KEYCLOAK_HOST')}/auth/realms/individual/protocol/openid-connect/logout?redirect_uri=${Env.get('SERVER_URL')}/login`);
-	}
-
 	async show ({ auth, params, view, response, request, session }) {
 		try {
 			// find user
@@ -406,98 +362,6 @@ class UserController {
 		});
 
 		return view.render('adminPages.viewUsers', { users, pageTitle, moment });
-	}
-
-	/**
-	 * Redirect to Active Directory login form login screen
-	 *
-	 */
-	async loginAD ({ response }) {
-		const authUri = oauth2.authorizationCode.authorizeURL({
-			redirect_uri: Env.get('KEYCLOAK_REDIRECT_URI'),
-			scope: Env.get('KEYCLOAK_SCOPES')
-		});
-
-		return response.redirect(authUri);
-	}
-
-	/**
-	 * Render login page
-	 *
-	 */
-	async loginRender ({ auth, view, response }) {
-		// present login to logged out users only
-		if (auth.user) {
-			return response.redirect('/');
-		} else {
-			var numb = Math.floor(Math.random() * 8) + 1;
-			var photoName = 'login_' + numb + '.jpg';
-			return view.render('auth.login', { photoName });
-		}
-	}
-	
-	/**
-	 * Convert code to token to authenticate user
-	 * We return here after user logs in with AD
-	 */
-	async authAD ({ request, session, response, auth, view }) {
-		// getting authorization code
-		const code = request.only(['code']).code;
-
-		// acquiring access token for user
-		let userInfo;
-		if (code) {
-			try {
-				let result = await oauth2.authorizationCode.getToken({
-					code: code,
-					redirect_uri: Env.get('KEYCLOAK_REDIRECT_URI'),
-					scope: Env.get('KEYCLOAK_SCOPES')
-				});
-				const token = await oauth2.accessToken.create(result);
-
-				userInfo = JWT.decode(token.token.id_token);
-			} catch (err) {
-				console.log(err)
-				return err;
-			}
-		}
-
-		const email = userInfo.email;
-		const user = await User
-			.query()
-			.where('email', email.toLowerCase())
-			.first();
-
-		if (user) {
-			await auth.login(user);
-			if (auth.user.role_id === 2) {
-				if (auth.user.verified) {
-					session.flash({
-						notification: 'Welcome! You are logged in'
-					});
-
-					return response.redirect('/userDash');
-				} else {
-					return response.redirect('/profile');
-				}
-			} else {
-				return response.redirect('/');
-			}
-		} else {
-			let newUser = {};
-			newUser.firstname = userInfo.given_name;
-			newUser.lastname = userInfo.family_name;
-			newUser.email = userInfo.email.toLowerCase();
-			newUser.role_id = await UserRole.getRoleID('user');
-			newUser.verified = false;
-			newUser.building_id = '1';
-			newUser.tower_id = '1';
-			newUser.floor_id = '1';
-
-			const user = await User.create(newUser);
-			await auth.login(user);
-			return response.redirect('/profile');
-		}
 	}
 }
